Upstream 6.02 parity: FLASH LOAD IMAGE — sub-step 6a of TILEMAP/SPRITE port
Adds FLASH LOAD IMAGE slot, file$ [,O/OVERWRITE] to cmd_flash. Decodes
24-bit BMP from SD into the named flash slot, packed as RGB121 nibbles
(2 pixels per byte) preceded by an 8-byte [width:uint32 LE][height:
uint32 LE] header — the exact storage layout TILEMAP / TILEMAP SPRITE's
blit121 (sub-step 6b) will read from. Self-contained reader avoids
pulling in upstream's general decodeBMP / linecallback machinery, which
has a different shape from our BMP_bDecode (per-pixel DrawPixel)
decoder.

Ported (paraphrased) from upstream FileIO.c:947 + the flashpackline /
writetoflashcallback helpers in upstream FileIO.c. Even-indexed pixel
goes in the low nibble, odd in the high — matches upstream's packing
convention so blit121 reads correctly.

Host infra fixes uncovered by the new test:

- host_fs_shims.c: host_flash_target_buf was 4 KB and disjoint from
  flash_prog_buf, so flash_range_erase / flash_range_program writes to
  flash slots silently no-op'd (off >= HOST_FLASH_SIZE returned). Now
  sized to MAXFLASHSLOTS * MAX_PROG_SIZE and translated by offset so
  writes land in the right slot. Previously every FLASH SAVE / FLASH
  DISK LOAD on host was broken — no test exercised it.

- configuration.h: cap MAX_PROG_SIZE at 256 KB on wasm. The wasm build
  was overriding HEAP_MEMORY_SIZE to 8 MB so big BASIC programs could
  use lots of variables, which (via the MAX_PROG_SIZE = HEAP_MEMORY_SIZE
  conflation) ballooned the slot mirror to 24 MB and overflowed wasm's
  32 MB initial-memory ceiling. flash_prog_buf in host_wasm_main.c is
  already 256 KB, so programs >256 KB couldn't be saved through
  flash_range_program regardless. AllMemory[] still gets the full 8 MB
  for runtime variables.

- host_peripheral_stubs.c: fun_info was a stripped-down stub that
  returned 0 for MM.INFO(FLASH ADDRESS slot); fun_peek was a no-op.
  Both required for flash-slot inspection from BASIC. Implemented FLASH
  ADDRESS plus PEEK(INT8/BYTE/WORD/SHORT/INTEGER/FLOAT addr) using
  uintptr_t round-trips to avoid 64-bit pointer truncation.

- MM_Misc.c: same uintptr_t fix on the device path's MM.INFO(FLASH
  ADDRESS) so the cast is portable.

Test t200_flash_load_image.bas authors a 4×2 24-bit BMP byte-by-byte
through OPEN ... FOR OUTPUT + PRINT #fnbr, CHR\$(b);, runs FLASH LOAD
IMAGE 1, then PEEKs the 12 expected output bytes (4-byte width +
4-byte height + 4 packed RGB121 bytes) at MM.INFO(FLASH ADDRESS 1).
Interp-only because cmd_flash is bridged.

All four gates green: host 234/234, rp2040 .uf2, rp2350 .uf2, wasm.

+++ MM_Misc.c
@@ -4046,7 +4046,10 @@ void MIPS16 fun_info(void){
             targ=T_INT;
             return;
         } else if((tp=checkstring(ep, (unsigned char *)"FLASH ADDRESS"))){
-            iret=(int64_t)(unsigned int)(flash_target_contents + (getint(tp,1,MAXFLASHSLOTS) - 1) * MAX_PROG_SIZE);
+            /* uintptr_t round-trip — `(unsigned int)` truncates pointers
+             * on 64-bit hosts (macOS / wasm64) and silently zeroes the
+             * upper bits, returning 0 to BASIC. */
+            iret=(int64_t)(uintptr_t)(flash_target_contents + (getint(tp,1,MAXFLASHSLOTS) - 1) * MAX_PROG_SIZE);
             targ=T_INT;
             return;
         } else if((tp=checkstring(ep, (unsigned char *)"FILESIZE"))){
